+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added GamepadButton to the InputManager and ButtonConfig - custom inputs can now define a gamepad button corresponding to the keyboard input
 * Added gamepad buttons to KeyHints for custom inputs as well as vanilla key overrides
 * Gamepad buttons can be defined in the mod settings if they are bound to a config
+* Refactored custom KeyHints into their own KeyHintManager and obsoleted the API in the GUIManager
 
 ## Version 2.3.7
 * Fixed translation of the custom skill raise message